Move known firewall service definitions to core/network/firewall
