+++ apiserver/common/crossmodel/crossmodel.go
@@ -15,8 +15,8 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/core/cache"
 	"github.com/juju/juju/core/crossmodel"
-	"github.com/juju/juju/core/firewall"
 	"github.com/juju/juju/core/life"
+	"github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/network"
 )
