+++ apiserver/common/crossmodel/interface.go
@@ -8,8 +8,8 @@ import (
 	"gopkg.in/macaroon.v2"
 
 	"github.com/juju/juju/core/crossmodel"
-	"github.com/juju/juju/core/firewall"
 	"github.com/juju/juju/core/network"
+	"github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/core/permission"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/state"
