+++ apiserver/common/crossmodel/state.go
@@ -8,7 +8,7 @@ import (
 	"github.com/juju/names/v4"
 
 	"github.com/juju/juju/core/crossmodel"
-	"github.com/juju/juju/core/firewall"
+	"github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/state"
 )
