+++ apiserver/facades/client/firewallrules/firewallrules_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/juju/juju/apiserver/facades/client/firewallrules"
 	"github.com/juju/juju/apiserver/params"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
-	"github.com/juju/juju/core/firewall"
+	"github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/state"
 	coretesting "github.com/juju/juju/testing"
 )
