+++ apiserver/facades/controller/crossmodelrelations/crossmodelrelations_test.go
@@ -26,8 +26,8 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
 	"github.com/juju/juju/core/crossmodel"
-	corefirewall "github.com/juju/juju/core/firewall"
 	"github.com/juju/juju/core/life"
+	corefirewall "github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/state"
