+++ apiserver/facades/controller/firewaller/firewaller.go
@@ -17,8 +17,8 @@ import (
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
 	"github.com/juju/juju/apiserver/params"
-	corefirewall "github.com/juju/juju/core/firewall"
 	"github.com/juju/juju/core/network"
+	corefirewall "github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/state"
 	"github.com/juju/juju/state/watcher"
