+++ apiserver/facades/controller/firewaller/mock_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/controller"
 	"github.com/juju/juju/core/crossmodel"
-	corefirewall "github.com/juju/juju/core/firewall"
 	"github.com/juju/juju/core/network"
+	corefirewall "github.com/juju/juju/core/network/firewall"
 	"github.com/juju/juju/core/status"
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/state"
