+++ state/firewallrules.go
@@ -11,7 +11,7 @@ import (
 	"gopkg.in/mgo.v2/bson"
 	"gopkg.in/mgo.v2/txn"
 
-	"github.com/juju/juju/core/firewall"
+	"github.com/juju/juju/core/network/firewall"
 )
 
 // FirewallRule instances describe the ingress networks
